+++ content/en/UI.json
@@ -61,6 +61,7 @@
     "Blog": "Blog"
   },
   "Blog": "Blog",
+  "Blog_Language_Selector":  "Blog Language Selector",
   "Editorial": "Editorial",
     "Community": "Community",
     "Governance": "Governance",
