+++ src/modules/localization/LanguageSelector.js
@@ -1,6 +1,7 @@
 /** @jsx jsx */
 import { Fragment } from "react";
-import Select, { components } from "react-select";
+
+import { Select } from "@modules/ui";
 import { useNavigate } from "@reach/router";
 import { Box, jsx, Text, useThemeUI } from "theme-ui";
 import { trackCustomEvent } from "gatsby-plugin-google-analytics";
@@ -31,81 +32,8 @@ const LanguageSelector = ({ data, pagePath }) => {
     navigate(value);
   };
 
-  const Menu = (props) => (
-    <components.Menu
-      {...props}
-      sx={{
-        marginTop: 0,
-        borderRadius: "12px",
-        border: "1px solid",
-        borderColor: "muted",
-        overflow: "hidden",
-      }}
-    />
-  );
 
-  const MenuList = (props) => (
-    <components.MenuList {...props} sx={{ py: 0, borderRadius: "12px" }} />
-  );
 
-  const Option = (props) => (
-    <components.Option
-      {...props}
-      sx={{ px: "18px", cursor: "pointer", fontSize: ["18px", "18px", 3] }}
-    />
-  );
-
-  const IndicatorsContainer = (props) => (
-    <Box
-      sx={{
-        width: "auto",
-        pr: 2,
-        "& > * >span": {
-          display: "none",
-        },
-        "& svg, &:hover svg": {
-          color: props.isDisabled ? "muted" : "backgroundAlt",
-        },
-      }}
-    >
-      <components.IndicatorsContainer {...props} />
-    </Box>
-  );
-
-  const ValueContainer = (props) => (
-    <components.ValueContainer
-      sx={{
-        fontFamily: "body",
-        color: "textMuted",
-        pl: 3,
-      }}
-      {...props}
-    />
-  );
-
-  const Control = (props) => (
-    <components.Control
-      sx={{
-        border: "1px solid",
-        borderColor: "muted",
-        borderRadius: "12px",
-        fontSize: ["18px", "18px", 3],
-        pr: 0,
-        p: ["2px", "2px", 0],
-      }}
-      {...props}
-    />
-  );
-
-  //Override select component theme with our theme since it's not connected to theme-ui
-  const uiSelectTheme = {
-    primary: theme.colors.primary,
-    primary75: theme.colors.success,
-    primary50: theme.colors.primaryMuted,
-    primary25: theme.colors.successAlt,
-    danger: theme.colors.bear,
-    dangerLight: theme.colors.bearAlt,
-  };
 
   //If we have existing languages or we're swapping, show the select.
   if (data.length > 0) {
@@ -121,19 +49,6 @@ const LanguageSelector = ({ data, pagePath }) => {
       >
         <Select
           isSearchable={false}
-          theme={(selectTheme) => ({
-            ...selectTheme,
-            fontFamily: theme.fonts.body,
-            colors: { ...selectTheme.colors, ...uiSelectTheme },
-          })}
-          components={{
-            Menu,
-            MenuList,
-            IndicatorsContainer,
-            ValueContainer,
-            Control,
-            Option,
-          }}
           options={data}
           onChange={onChange}
           aria-label={t("Page_Language_Selector")}
