+++ src/modules/ui/index.js
@@ -13,3 +13,4 @@ export { default as Accordion } from "./Accordion";
 export { default as Checklist } from "./Checklist";
 export { default as Column } from "./Column";
 export { default as Code } from "./Code";
+export {default as  Select} from './Select'
